Move logging functionality to a separate file
